pyproject.toml
[tool.poetry]
name = "hartree-fork"
version = "0.1.0"
description = "Hartree Fock, a Self Consistent Field (SCF) method"
authors = ["RenChu Wang <patrick1031wang@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "hartree_fork"}]

[tool.poetry.dependencies]
hydra-core = "^1.3.2"
numba = "^0.56.4"
numpy = ">=1.23,<1.24"
omegaconf = "^2.3.0"
python = ">=3.10,<3.12"
rich = "^13.3.4"
scipy = "^1.10.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
